Changing GPIO read/write values from u32 to u8

+++ 01-MCAL/02-GPIO_DRIVER/GPIO.c
@@ -126,7 +126,7 @@ extern ERROR_STATUS GPIO_writePin(GPIO_t * peri, u32 value)
   Output: ERROR_STATUS
 
  */
-extern ERROR_STATUS GPIO_directWritePin(void * port ,u32 pin, u32 value)
+extern ERROR_STATUS GPIO_directWritePin(void * port ,u32 pin, u8 value)
 {
   ERROR_STATUS currentStatus = status_Ok;
   
@@ -186,7 +186,7 @@ extern ERROR_STATUS GPIO_readPin(GPIO_t * peri, u32 *value)
   Output: ERROR_STATUS
 
  */
-extern ERROR_STATUS GPIO_directReadPin(void * port ,u32 pin, u32 * value)
+extern ERROR_STATUS GPIO_directReadPin(void * port ,u32 pin, u8 * value)
 {
   ERROR_STATUS status = status_Ok;
